Add functional buttons for playerSelection
Buttons were added to the HTML to allow a player to choose rock, paper
, or scissors for the game. The event listener initiates the game so
that no rounds will begin until the player makes a choice.

Commented out the for loop and finalResult function to get functionality
working for the buttons only and so I can find a way to include this
logic to make the game play 5 rounds once more.